Refactor the error handling code
Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:
