+++ db/db_impl.h
@@ -68,44 +68,6 @@ struct JobContext;
 struct ExternalSstFileInfo;
 struct MemTableInfo;
 
-class ErrorHandler {
-  public:
-    ErrorHandler(const ImmutableDBOptions& db_options,
-        InstrumentedMutex* db_mutex)
-      : db_options_(db_options),
-        bg_error_(Status::OK()),
-        db_mutex_(db_mutex)
-      {}
-    ~ErrorHandler() {}
-
-    Status::Severity GetErrorSeverity(BackgroundErrorReason reason,
-        Status::Code code, Status::SubCode subcode);
-
-    Status SetBGError(const Status& bg_err, BackgroundErrorReason reason);
-
-    Status GetBGError()
-    {
-      return bg_error_;
-    }
-
-    void ClearBGError() {
-      bg_error_ = Status::OK();
-    }
-
-    bool IsDBStopped() {
-      return !bg_error_.ok();
-    }
-
-    bool IsBGWorkStopped() {
-      return !bg_error_.ok();
-    }
-
-  private:
-    const ImmutableDBOptions& db_options_;
-    Status bg_error_;
-    InstrumentedMutex* db_mutex_;
-};
-
 class DBImpl : public DB {
  public:
   DBImpl(const DBOptions& options, const std::string& dbname,
