+++ db/db_impl_debug.cc
@@ -10,6 +10,7 @@
 #ifndef NDEBUG
 
 #include "db/db_impl.h"
+#include "db/error_handler.h"
 #include "monitoring/thread_status_updater.h"
 
 namespace rocksdb {
