+++ src.mk
@@ -24,8 +24,8 @@ LIB_SOURCES =                                                   \
   db/db_impl_write.cc                                           \
   db/db_info_dumper.cc                                          \
   db/db_iter.cc                                                 \
-  db/db_error_handler.cc					\
   db/dbformat.cc                                                \
+  db/error_handler.cc						\
   db/event_helpers.cc                                           \
   db/experimental.cc                                            \
   db/external_sst_file_ingestion_job.cc                         \
@@ -270,7 +270,6 @@ MAIN_SOURCES =                                                          \
   db/db_compaction_test.cc                                              \
   db/db_dynamic_level_test.cc                                           \
   db/db_encryption_test.cc                                              \
-  db/db_error_handling_test.cc                                        	\
   db/db_flush_test.cc                                                   \
   db/db_inplace_update_test.cc                                          \
   db/db_io_failure_test.cc                                              \
@@ -295,6 +294,7 @@ MAIN_SOURCES =                                                          \
   db/dbformat_test.cc                                                   \
   db/deletefile_test.cc                                                 \
   db/env_timed_test.cc                                                  \
+  db/error_handler_test.cc                                        	\
   db/external_sst_file_basic_test.cc                                    \
   db/external_sst_file_test.cc                                          \
   db/fault_injection_test.cc                                            \
